reorganizing api to pong
