+++ srcs/frontend/static/js/api/index.js
@@ -1,3 +1,4 @@
 import Friends from "./friends.js";
+import PongData from "./pongData.js";
 
-export { Friends };
+export { Friends, PongData };
