+++ srcs/frontend/static/js/services/index.js
@@ -1,6 +1,5 @@
 import i18nextInit from "./i18next.js";
 import renderPage from "./render.js";
 import renderSidebar from "./renderSidebar.js";
-import FetchData from "./FetchData.js";
 
-export { i18nextInit, renderPage, renderSidebar, FetchData };
+export { i18nextInit, renderPage, renderSidebar };
